Update and rename tkinter.py to minimal_tkinter.py
